+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
+### Added
+- Implement `DbcLibrary`
+
 ### Changed
 - Update and clean out old deps
 - Replace regex parsing with `nom` to handle strings with newlines